feat: updates the export UI to remove multiple rounds
